chore: Annotate types of params, use ANY

+++ backend/sql/reservations.sql
@@ -6,5 +6,5 @@ SELECT * FROM reservations
 
 -- name: AddReservation :one
 INSERT INTO reservations (flight_id, firstname, lastname, email, reservation_datetime, status)
-VALUES (@flight_id, @firstname, @lastname, @email, NOW(), @status)
+VALUES (@flight_id::int, @firstname, @lastname, @email, NOW(), @status)
 RETURNING *;
